refactor: re-organiza arquivos de cards
