+++ src/features/acf/diabetes/backend/internalCards/repository.ts
@@ -1,8 +1,10 @@
 import { prisma } from "@prisma/prismaClient";
-import { DiabetesAcfInternalCardsHealthIndicator, type InternalCardDBDataItem } from "./internalCards.model";
+import { DiabetesAcfInternalCardsHealthIndicator, type InternalCardDBDataItem } from "../model";
 
 export type MunicipalitySusIdAndTeamIne = {
+    // biome-ignore lint/style/useNamingConvention: <explanation>
     municipio_id_sus: string;
+    // biome-ignore lint/style/useNamingConvention: <explanation>
     equipe_ine_cadastro: string;
 }
 
@@ -18,7 +20,9 @@ export const totalPatientsWithExamsAndAppointment = async (municipalitySusIdAndT
     return await prisma.impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos.count({
         where: {
             ...municipalitySusIdAndTeamIne,
+            // biome-ignore lint/style/useNamingConvention: <explanation>
             prazo_proxima_consulta: "Em dia",
+            // biome-ignore lint/style/useNamingConvention: <explanation>
             prazo_proxima_solicitacao_hemoglobina: "Em dia"
         },
     })
@@ -28,6 +32,7 @@ export const totalPatientsSelfDiagnosed = async (municipalitySusIdAndTeamIne: Mu
     return await prisma.impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos.count({
         where: {
             ...municipalitySusIdAndTeamIne,
+            // biome-ignore lint/style/useNamingConvention: <explanation>
             identificacao_condicao_diabetes: "Autorreferida"
         },
     })
@@ -37,6 +42,7 @@ export const totalPatientsWithClinicalDiagnosis = async (municipalitySusIdAndTea
     return await prisma.impulso_previne_dados_nominais___painel_enfermeiras_lista_nominal_diabeticos.count({
         where: {
             ...municipalitySusIdAndTeamIne,
+            // biome-ignore lint/style/useNamingConvention: <explanation>
             identificacao_condicao_diabetes: "Diagnóstico Clínico"
         },
     })
@@ -47,7 +53,9 @@ export const internalCardsDataForTeam = async(
     municipalitySusID: string,
 ): Promise<readonly InternalCardDBDataItem[]> =>{
     const municipalitySusIDAndTeamIne = {
+        // biome-ignore lint/style/useNamingConvention: <explanation>
         municipio_id_sus: municipalitySusID,
+        // biome-ignore lint/style/useNamingConvention: <explanation>
         equipe_ine_cadastro: teamIne
     }
