+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/Internalcards/presentation.tsx
@@ -2,7 +2,7 @@
 import { CardGrid } from '@impulsogov/design-system';
 import type { CardProps } from '@impulsogov/design-system/dist/molecules/Card/Card';
 
-export const CardsGridClientWrapper = ({internalCardsProps}:{internalCardsProps : CardProps[]}) => (
+export const InternalCards = ({internalCardsProps}:{internalCardsProps : CardProps[]}) => (
     <div style={{marginTop: '75px'}}>
         <CardGrid cards={internalCardsProps} />
     </div>
