--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/cards/cardsDetails.ts
@@ -1,20 +1,0 @@
-import type { CardDetailsMap } from "@/helpers/cardsList";
-
-export const cardsDetails: CardDetailsMap = {
-    INDICADOR_1: {
-        title: "Indicador 1",
-        titlePosition: "top",
-    },
-    INDICADOR_2: {
-        title: "Indicador 2",
-        titlePosition: "top",
-    },
-    INDICADOR_3: {
-        title: "Indicador 3",
-        titlePosition: "top",
-    },
-    INDICADOR_4: {
-        title: "Indicador 4",
-        titlePosition: "top",
-    },
-};
