--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/cards/externalCards/externalCards.adapter.ts
@@ -1,13 +1,0 @@
-import type { AcfDashboardType } from "../../../../../../../../../common/model"
-import type { DiabetesAcfExternalCardsDescriptionType, ExternalCardDBDataItem, ExternalCardDataItem } from "./externalCards.model"
-
-
-export const externalCardsDbToModel = (
-    data: ExternalCardDBDataItem[]
-): ExternalCardDataItem[] => {
-    return data.map((item) => ({
-        value: item.valor,
-        acfDashboardType: item.lista as AcfDashboardType,
-        acfExternalCardsDescription: item.descricao as DiabetesAcfExternalCardsDescriptionType
-    }))
-}
