--- src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/dashboards/modules/cards/externalCards/externalCards.model.ts
@@ -1,22 +1,0 @@
-import type { AcfDashboardType } from "../../../../../../../../../common/model";
-
-export type DiabetesAcfExternalCardsDescriptionType = "DIAGNOSTICO_AUTORREFERIDO" | "DIAGNOSTICO_CLINICO"
-
-export type ExternalCardItem = ExternalCardDataItem &{
-    municipalitySusID: string,
-    teamIne: string,
-}
-
-export type ExternalCardDataItem = {
-    acfDashboardType: AcfDashboardType,
-    acfExternalCardsDescription: DiabetesAcfExternalCardsDescriptionType;
-    value: number;
-} 
-
-export type ExternalCardDBDataItem = {
-    municipio_id_sus: string;
-    ine: string;
-    lista: string;
-    valor: number;
-    descricao: string;
-}
